Move all deeplink classes to a separate package
