+++ WordPress/src/main/java/org/wordpress/android/modules/ViewModelModule.java
@@ -3,7 +3,7 @@ package org.wordpress.android.modules;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-import org.wordpress.android.ui.DeepLinkingIntentReceiverViewModel;
+import org.wordpress.android.ui.deeplinks.DeepLinkingIntentReceiverViewModel;
 import org.wordpress.android.ui.JetpackRemoteInstallViewModel;
 import org.wordpress.android.ui.accounts.LoginViewModel;
 import org.wordpress.android.ui.accounts.login.LoginNoSitesErrorViewModel;
