+++ WordPress/src/main/java/org/wordpress/android/ui/deeplinks/ServerTrackingHandler.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui
+package org.wordpress.android.ui.deeplinks
 
 import com.android.volley.Request.Method
 import com.android.volley.RequestQueue
