+++ WordPress/src/test/java/org/wordpress/android/ui/deeplinks/DeepLinkUriUtilsTest.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui
+package org.wordpress.android.ui.deeplinks
 
 import com.nhaarman.mockitokotlin2.whenever
 import org.assertj.core.api.Assertions.assertThat
