+++ WordPress/src/test/java/org/wordpress/android/ui/deeplinks/UriTestHelper.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui
+package org.wordpress.android.ui.deeplinks
 
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
